Initial version of the discovery api
Discovery API is able to parse nmap files and map open ports to targeted services.
Access tests work for mongoDB with basic authentication testing (username,password = admin) but redis access test still needs to be refined because of a problem with pexpect.
Modular structure achieved with json templates that specify commands to send to the client and answers to expect. However, this requires downloading the specific clients, in this case the redis and the mongoDB client. python libraries for both exist but this would defeat the modular structure of the equaeris scanner.
Might however change the general design if the modular structure using clients seams unfeasible